Create a tester object along with the custom chain.
The tester object is used to record information as the test is being
executed that can later be checked against the final state of the
chain.

In this version it only records the blocks by name. This allows us to
check for example if the block chain ends up with the correct top
block.