Fixed memory leak in mineNextBlock.
All previous blocks were being retained in memory because prev_hash
contained a reference to the buffer from hash in the previous
block. And that in turn has a reference of the block object, so no
block could be freed. By creating a fresh copy of the hash buffer we
avoid the issue.