Major refactoring of block chain download.
The new block chain works by sorting incoming orphan blocks into a
series of linked lists. These can then be discarded based on a
timeout.

We also now maintain a cache of recently added blocks, because chain
splits and duplicate downloads are likely to concern recently
downloaded blocks.

This commit is incomplete, there are still a good number of TODOs and
testing left to be done, but it does seem to work at least for the
single node download and non-evil p2p download cases. It also provides
a much better foundation for future bugfixes.