Removed transactions' awareness of their own position in the chain.
BitcoinJS stores information about side chains, which can lead to the
same transaction appearing twice in the database. So the schema for
the transaction mustn't contain a reference to the block (since there
could be multiple blocks.)

Note that this commit also removes the "active" flag from
transactions. This flag wasn't used and also wasn't properly updated,
but in the future we may want to re-add it in some form to speed up
connecting a transaction to its inputs.