Fully connect all inputs during block chain download.
This patch contains quite a bit of refactoring as well, which was
necessary to make this work.

Note we don't actually run the scripts yet during download (we do for
memory pool acceptance). The reason is that we need an optimization
where the verification runs in its own thread, otherwise the download
is simply too slow.

This patch also fixes the indexing of coinbase and IP transactions.

The test suite now runs each block chain test case twice - once to
generate the test data and once to simulate a download as it would
happen in reality.